backend/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
google-generativeai
langchain
langchain-google-genai==2.0.10
python-dotenv==1.1.1
dj-database-url
psycopg2-binary
gunicorn
whitenoise
xhtml2pdf
django-cors-headers==4.3.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
mongoengine==0.29.1
pymongo==4.10.1
dnspython==2.6.1
PyMuPDF==1.26.5
python-docx==1.2.0
markdown
cloudinary
PyPDF2
reportlab
